python/memory-sample-agent/pyproject.toml
[project]
name = "customer-service-agent-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "teams_memory",
    "python-dotenv",
    "aiohttp",
    "teams-ai",
    "botbuilder",
    "pydantic>=2.10.3",
    "litellm==1.54.1",
]

[tool.uv.sources]
teams_memory = { workspace = true }
